Refactor dropdown handling code
