+++ coursework/src/app/recipes/recipe-detail/recipe-detail.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 
+import { DropdownHandler } from '../../shared/dropdown-handler.class';
 import { Recipe } from '../recipe.model';
 
 @Component({
@@ -7,33 +8,17 @@ import { Recipe } from '../recipe.model';
     templateUrl: './recipe-detail.component.html',
     styleUrls: ['./recipe-detail.component.css']
 })
-export class RecipeDetailComponent implements OnInit {
-    dropdownOpen: boolean = false;
-    dropdownId: string = 'recipe-detail-dropdown';
-
+export class RecipeDetailComponent extends DropdownHandler implements OnInit {
     @Input('recipeData') recipe: Recipe;
 
-    constructor() { }
+    constructor() {
+        super('recipe-detail-dropdown');
+    }
 
     ngOnInit(): void {
     }
 
     handleOutsideDropdownClick(e: any) {
-        // There may be other dropdowns...
-        const isCorrectDropdown = () => {
-            if ('dropdownId' in e.target.dataset) {
-                return (this.dropdownId === e.target.dataset.dropdownId);
-            }
-            return false;
-        };
-
-        // Will close the dropdown if clicked outside of it when open
-        if (this.dropdownOpen && !e.target.className.includes('dropdown')) {
-            this.dropdownOpen = false;
-        }
-        else if (this.dropdownOpen && e.target.className.includes('dropdown') && !isCorrectDropdown()) {
-            // Will close the dropdown when clicking a different dropdown
-            this.dropdownOpen = false;
-        }
+        return this._handleOutsideDropdownClick(e);
     }
 }
